% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_number_int.R
\name{cut_number_int}
\alias{cut_number_int}
\title{Cut integerish data nicely}
\usage{
cut_number_int(x, n, retry = 3L, show_highest_value = FALSE)
}
\arguments{
\item{x}{a numeric vector which is to be converted to a factor by cutting.}

\item{n}{number of bins to create.}

\item{retry}{number of maximum attempts to calculate the non-duplicated breaks for the given.}

\item{show_highest_value}{if \code{FALSE}, do not show the highest value in the label (i.e. the label would be \code{"100~"}).}
}
\description{
Cut integerish data nicely
}
