% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutnumberint-package.R
\docType{package}
\name{cutnumberint-package}
\alias{cutnumberint}
\alias{cutnumberint-package}
\title{cutnumberint: Cut integer-ish data nicely}
\description{
Provides a nicer version of cut_number()
}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gmail.com} (\href{https://orcid.org/0000-0002-3385-7233}{ORCID})

}
\keyword{internal}
